+++ packages/core/src/__tests__/makeRandomAddress.ts
@@ -1,4 +1,4 @@
-import * as ecc from "tiny-secp256k1";
+import ecc from "@bitcoinerlab/secp256k1";
 import ECPairFactory from "ecpair";
 import * as bitcoin from "bitcoinjs-lib";
 import { extractXCoordinate } from "~/utils";