+++ packages/core/src/actions/getUTXOs.test.ts
@@ -1,6 +1,6 @@
 import * as bitcoin from "bitcoinjs-lib";
 import ECPairFactory from "ecpair";
-import * as ecc from "tiny-secp256k1";
+import ecc from "@bitcoinerlab/secp256k1";
 import {
 	afterEach,
 	assert,