+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @midl-xyz/midl-js-docs
 
+## 0.0.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.36
+  - @midl-xyz/midl-js-react@0.0.43
+
+## 0.0.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.35
+  - @midl-xyz/midl-js-react@0.0.42
+
 ## 0.0.41
 
 ### Patch Changes