+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,44 @@
 # e2e
 
+## 1.0.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.72
+
+## 1.0.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.71
+
+## 1.0.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.70
+
+## 1.0.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.36
+  - @midl-xyz/midl-js-executor@0.0.69
+  - @midl-xyz/midl-js-react@0.0.43
+
+## 1.0.40
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@0.0.68
+  - @midl-xyz/midl-js-core@0.0.35
+  - @midl-xyz/midl-js-react@0.0.42
+
 ## 1.0.39
 
 ### Patch Changes