+++ apps/rune-etcher/CHANGELOG.md
@@ -1,5 +1,21 @@
 # rune-etcher
 
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.36
+  - @midl-xyz/midl-js-react@0.0.43
+
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.35
+  - @midl-xyz/midl-js-react@0.0.42
+
 ## 0.1.23
 
 ### Patch Changes