+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @midl-xyz/midl-js-core
 
+## 0.0.36
+
+### Patch Changes
+
+- feat: add keypair connector
+
+## 0.0.35
+
+### Patch Changes
+
+- feat: migrate ecc to bitcoinerlab
+
 ## 0.0.34
 
 ### Patch Changes